+++ R/LearnerClustDBSCANfpc.R
@@ -69,6 +69,11 @@ LearnerClustDBSCANfpc = R6Class(
       if (isFALSE(self$param_set$values$seeds)) {
         error_config("Predicting requires seed points, train with `seeds = TRUE`.")
       }
+      if (isTRUE(self$param_set$values$scale)) {
+        error_config(
+          "Predicting is not supported for `scale = TRUE` since `fpc:::predict.dbscan()` ignores the scaling."
+        )
+      }
       partition = as.integer(invoke(
         predict,
         self$model,