+++ tests/testthat/test_mlr_learners_clust_pam.R
@@ -7,6 +7,13 @@ test_that("autotest", {
   expect_true(result, info = result$error)
 })
 
+test_that("predicting errors informatively for unsupported training options", {
+  task = tsk("usarrests")
+  learner = lrn("clust.pam", stand = TRUE)
+  learner$train(task)
+  expect_snapshot(error = TRUE, learner$predict(task))
+})
+
 test_that("Learner properties are respected", {
   task = tsk("usarrests")
   learner = lrn("clust.pam")